feat validate generalisation
